+++ packages/firebase_auth/firebase_auth/ios/Classes/Private/PigeonParser.h
@@ -4,7 +4,7 @@
  * BSD-style license that can be found in the LICENSE file.
  */
 
-#import <Firebase/Firebase.h>
+@import FirebaseAuth;
 #import <Foundation/Foundation.h>
 #import "firebase_auth_messages.g.h"