+++ packages/firebase_data_connect/firebase_data_connect/lib/src/core/empty_serializer.dart
@@ -1,0 +1,8 @@
+// Copyright 2024, the Chromium project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+// Empty serializer to be used when a null variable is passed.
+String emptySerializer(void _) {
+  return '';
+}