+++ packages/firebase_storage/firebase_storage/ios/Classes/FLTTaskStateChannelStreamHandler.h
@@ -10,7 +10,7 @@
 #import <Flutter/Flutter.h>
 #endif
 
-#import <Firebase/Firebase.h>
+@import FirebaseStorage;
 
 #import <Foundation/Foundation.h>