+++ packages/firebase_database/firebase_database_platform_interface/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.5+44
+
+ - Update a dependency to the latest release.
+
+## 0.2.5+43
+
+ - Update a dependency to the latest release.
+
 ## 0.2.5+42
 
  - Update a dependency to the latest release.