+++ packages/firebase_in_app_messaging/firebase_in_app_messaging/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.8.0+8
+
+ - Update a dependency to the latest release.
+
+## 0.8.0+7
+
+ - Update a dependency to the latest release.
+
 ## 0.8.0+6
 
  - Update a dependency to the latest release.