+++ packages/firebase_in_app_messaging/firebase_in_app_messaging_platform_interface/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.4+44
+
+ - Update a dependency to the latest release.
+
+## 0.2.4+43
+
+ - Update a dependency to the latest release.
+
 ## 0.2.4+42
 
  - Update a dependency to the latest release.