+++ packages/firebase_storage/firebase_storage/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 12.3.2
+
+ - **FIX**(all,apple): use modular headers to import ([#13400](https://github.com/firebase/flutterfire/issues/13400)). ([d7d2d4b9](https://github.com/firebase/flutterfire/commit/d7d2d4b93e7c00226027fffde46699f3d5388a41))
+
+## 12.3.1
+
+ - Update a dependency to the latest release.
+
 ## 12.3.0
 
  - **FEAT**(web): update to `web: ^1.0.0` ([#13200](https://github.com/firebase/flutterfire/issues/13200)). ([8fab04ae](https://github.com/firebase/flutterfire/commit/8fab04aec3b95789856d95639131bf09db29175b))